refactor: rename docker-images to images
