--- src/components/molecules/DockerImageDetails/index.ts
@@ -1,1 +1,0 @@
-export {default} from './DockerImageDetails';
