+++ src/components/molecules/ImageDetails/ImageDetails.styled.tsx
