+++ src/components/molecules/ImageDetails/index.ts
@@ -1,0 +1,1 @@
+export {default} from './ImageDetails';
